Remove unstable integration test
* Removed unstable proposal API integration test
  * Test will be added in another PR
  * This instability is causing tests to fail in travis and blocking PRs
* Fixed inconsistent file naming in `/tests/integration/server`
* Removed duplicate file in `/tests/integration/server`
* A module level setup funciton in `/tests/integration/server/roles_api_test.py`
  was improperly set as a test function with no assertion.

[ Ticket: #364 ]
