+++ tests/integration/server/proposals_api_test.py
@@ -12,47 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # -----------------------------------------------------------------------------
-"""Integration tests for role APIs"""
-import requests
+"""Integration tests for proposals APIs"""
 
-
-from tests.utilities import (
-    create_test_user,
-    delete_user_by_username,
-    delete_role_by_name,
-    insert_role,
-)
-
-
-def test_proposals():
-    """Create a new fake role and try to add yourself to role you created"""
-    with requests.Session() as session:
-        user_payload = {
-            "name": "Susan S",
-            "username": "susans2224",
-            "password": "12345678",
-            "email": "susans@biz.co",
-        }
-        user_response = create_test_user(session, user_payload)
-        user_id = user_response.json()["data"]["user"]["id"]
-        role_resource = {
-            "name": "Office_Assistant",
-            "owners": user_id,
-            "administrators": user_id,
-        }
-        insert_role(role_resource)
-        delete_role_by_name("Office_Assistant")
-        role_response = session.post(
-            "http://rbac-server:8000/api/roles", json=role_resource
-        )
-        role_id = role_response.json()["data"]["id"]
-        insert_role(role_resource)
-        res = session.post(
-            "http://rbac-server:8000/api/roles/" + role_id + "/members",
-            json=user_response.json()["data"]["user"],
-        )
-        assert (
-            res.json()["message"] == "Owner is the requester. Proposal is autoapproved"
-        )
-        delete_user_by_username("susans2224")
-        delete_role_by_name("Office_Assistant")
+# TODO: Add tests for the proposals API
+# TODO: Add a test to check that users from the manager chain are in the
+#      approvers list.
