+++ tests/integration/server/roles_api_test.py
@@ -30,7 +30,7 @@ from tests.utilities import (
 from tests.rbac.api.assertions import assert_api_success
 
 
-def test_create_unique_role():
+def setup_module():
     """Create a new fake role resource which is unique"""
     with requests.Session() as session:
         user_payload = {
@@ -46,7 +46,41 @@ def test_create_unique_role():
             "owners": user_id,
             "administrators": user_id,
         }
-        response = session.post("http://rbac-server:8000/api/roles", json=role_resource)
+        session.post("http://rbac-server:8000/api/roles", json=role_resource)
+
+
+def test_proposals():
+    """Create a new fake role and try to add yourself to role you created"""
+    with requests.Session() as session:
+        user_payload = {
+            "name": "Susan S",
+            "username": "susans2224",
+            "password": "12345678",
+            "email": "susans@biz.co",
+        }
+        user_response = create_test_user(session, user_payload)
+        user_id = user_response.json()["data"]["user"]["id"]
+        role_resource = {
+            "name": "Office_Assistant",
+            "owners": user_id,
+            "administrators": user_id,
+        }
+        insert_role(role_resource)
+        delete_role_by_name("Office_Assistant")
+        role_response = session.post(
+            "http://rbac-server:8000/api/roles", json=role_resource
+        )
+        role_id = role_response.json()["data"]["id"]
+        insert_role(role_resource)
+        res = session.post(
+            "http://rbac-server:8000/api/roles/" + role_id + "/members",
+            json=user_response.json()["data"]["user"],
+        )
+        assert (
+            res.json()["message"] == "Owner is the requester. Proposal is autoapproved"
+        )
+        delete_user_by_username("susans2224")
+        delete_role_by_name("Office_Assistant")
 
 
 def test_create_duplicate_role():
